+++ lib/block-supports/elements.php
@@ -18,7 +18,7 @@ function gutenberg_render_elements_support( $block_content, $block ) {
 	/*
 	* For now we only care about link color.
 	* This code in the future when we have a public API
-	* should take advantage of WP_Theme_JSON::compute_style_properties
+	* should take advantage of WP_Theme_JSON_Gutenberg::compute_style_properties
 	* and work for any element and style.
 	*/
 	if ( null === $link_color ) {