+++ packages/components/src/ui/shortcut/component.tsx
@@ -10,6 +10,7 @@ import type { Ref } from 'react';
 import { useContextSystem, contextConnect } from '../context';
 // eslint-disable-next-line no-duplicate-imports
 import type { PolymorphicComponentProps } from '../context';
+import { View } from '../../view';
 
 export interface ShortcutDescription {
 	display: string;
@@ -25,10 +26,12 @@ function Shortcut(
 	props: PolymorphicComponentProps< Props, 'span' >,
 	forwardedRef: Ref< any >
 ): JSX.Element | null {
-	const { shortcut, className, ...otherProps } = useContextSystem(
-		props,
-		'Shortcut'
-	);
+	const {
+		as: asProp = 'span',
+		shortcut,
+		className,
+		...otherProps
+	} = useContextSystem( props, 'Shortcut' );
 	if ( ! shortcut ) {
 		return null;
 	}
@@ -44,14 +47,15 @@ function Shortcut(
 	}
 
 	return (
-		<span
+		<View
+			as={ asProp }
 			className={ className }
 			aria-label={ ariaLabel }
 			ref={ forwardedRef }
 			{ ...otherProps }
 		>
 			{ displayText }
-		</span>
+		</View>
 	);
 }